third_party/WebKit/LayoutTests/media/video-controls-show-on-focus.html

<!DOCTYPE html>
<title>Test visibiblity of controls when focusing video.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script src="media-controls.js"></script>
<video controls autoplay></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.onplaying = t.step_func(function() {
        runAfterHideMediaControlsTimerFired(t.step_func_done(function() {
            var controls = mediaControlsButton(video, "panel");
            assert_equals(getComputedStyle(controls).opacity, "0");
            video.focus();
            assert_equals(getComputedStyle(controls).opacity, "1");

            // We fade out when focus is lost, so the opacity is still 1
            video.blur();
            assert_equals(getComputedStyle(controls).opacity, "1");
        }), video);
    });

    video.src = findMediaFile("video", "content/test");
});
</script>
